# src/api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Petstore - OpenAPI 3.0",
    "description": "Just challenge to LITC",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/"
    }
  ],
  "paths": {
    "/api/product/market_name": {
      "get": {
        "summary": "Finds products by market",
        "description": "Finds products by market",
        "operationId": "get_products_by_market",
        "parameters": [
          {
            "name": "market_name",
            "in": "path",
            "description": "market name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/product"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/api/product": {
      "get": {
        "summary": "Get all products",
        "description": "Get all products",
        "operationId": "get_all_products",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new product",
        "description": "Add a new product",
        "operationId": "create_product",
        "requestBody": {
          "description": "Add a new product",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/product"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/product"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/product"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "summary": "Update an existing product",
        "description": "Update an existing product",
        "operationId": "edit_product",
        "requestBody": {
          "description": "Update an existing product",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/product"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/product"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Product not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/market": {
      "post": {
        "summary": "Create new market",
        "description": "Create new market",
        "operationId": "create_market",
        "requestBody": {
          "description": "Create new market",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/market"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/market"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "summary": "Get all market names",
        "description": "Get all market names",
        "operationId": "get_all_markets",
        "requestBody": {
          "description": "Get all market names",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/market"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/market"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "market": {
            "type": "string"
          },
          "calories": {
            "type": "integer"
          },
          "price": {
            "type": "integer"
          }
        }
      },
      "market": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      }
    }
  }
}